requirements.txt
requests>=2.31.0
PyGithub>=1.59.1
click>=8.1.7
colorama>=0.4.6
tabulate>=0.9.0
python-dotenv>=1.0.0